Move negatinp and sswell from subcomponent to component physics module, make negatinp an argument of PHYSICS instead of ZIEGER
