+++ rompy/swan/components/physics.py
@@ -1,7 +1,7 @@
 """Model physics components."""
 import logging
 from typing import Literal, Optional, Union
-from pydantic import root_validator, Field
+from pydantic import validator, root_validator, Field, confloat
 
 from rompy.swan.components.base import BaseComponent
 from rompy.swan.subcomponents.physics import (
@@ -14,9 +14,6 @@ from rompy.swan.subcomponents.physics import (
     ST6C3,
     ST6C4,
     ST6C5,
-    ROGERS,
-    ARDHUIN,
-    ZIEGER,
     WCAPKOMEN,
     WCAPAB,
     QUADRUPL,
@@ -45,6 +42,9 @@ SOURCE_TERMS = Union[
 ]
 
 
+#======================================================================================
+# Wave generation GEN1 | GEN2 | GEN3
+#======================================================================================
 class GEN1(BaseComponent):
     """Physics component GEN1.
 
@@ -175,6 +175,149 @@ class GEN3(BaseComponent):
         return repr
 
 
+#======================================================================================
+# Swell dissipation SSWELL
+#======================================================================================
+class NEGATINP(BaseComponent):
+    """Nonbreaking dissipation of Young et al. (2013) updated by Zieger et al. (2015).
+
+    With this optional command the user activates negative wind input. This is intended
+    only for use with non-breaking swell dissipation SSWELL ZIEGER. Parameter `rdcoef`
+    is a fraction between 0 and 1, representing the strength of negative wind input. As
+    an example, with [rdcoef]=0.04, for a spectral bin that is opposed to the wind
+    direction, the wind input factor W(k, θ) is negative, and its magnitude is 4% of
+    the corresponding value of the spectral bin that is in the opposite direction
+    (i.e. in the wind direction). See Zieger et al. (2015) eq. 11, where a0 is their
+    notation for [rdcoef]. Default [rdcoef]=0.0 and `rdcoef=0.04` is recommended,
+    though as implied by Zieger et al. (2015), this value is not well-established, so
+    the user is encouraged to experiment with other values.
+
+    References
+    ----------
+    Zieger, S., A.V. Babanin, W. E. Rogers and I.R. Young, 2015: Observation-based
+    source terms in the third-generation wave model WAVEWATCH, Ocean Model., 96, 2-25.
+
+    """
+
+    model_type: Literal["negatinp"] = Field(
+        default="negatinp", description="Model type discriminator"
+    )
+    rdcoef: confloat(ge=0.0, le=1.0) = Field(
+        default=0.0,
+        description="Coefficient representing the strength of negative wind input",
+    )
+
+    def cmd(self) -> str:
+        return f"NEGATINP rdcoef={self.rdcoef}"
+
+
+class SSWELL(BaseComponent):
+    """Swell dissipation sub-component.
+
+    `SSWELL base component.`
+
+    With this command the user can influence the type of swell dissipation included in
+    the computations. The Zieger option is intended for use with negative wind input
+    via the NEGATINP command. Zieger non-breaking dissipation follows the method used
+    in WAVEWATCH III version 4 and does not include the steepness-dependent swell
+    coefficient introduced in WAVEWATCH III version 5. As noted already, if
+    `GEN3 ST6...` command is used, this SSWELL command should be provided.
+
+    """
+
+    model_type: Literal["sswell"] = Field(
+        default="sswell", description="Model type discriminator"
+    )
+
+    def cmd(self) -> str:
+        return "SSWELL"
+
+
+class ROGERS(SSWELL):
+    """Nonbreaking dissipation of Rogers et al. (2012).
+
+    References
+    ----------
+    Rogers, W.E., A.V. Babanin and D.W. Wang, 2012: Observation-Consistent Input and
+    Whitecapping Dissipation in a Model for Wind-Generated Surface Waves: Description
+    and Simple Calculations, J. Atmos. Oceanic Technol., 29:9, 1329-1346.
+
+    """
+
+    model_type: Literal["rogers"] = Field(
+        default="rogers", description="Model type discriminator"
+    )
+    cdsv: Optional[float] = Field(
+        description="Coefficient related to laminar atmospheric boundary layer (SWAN default: 1.2)"
+    )
+    feswell: Optional[float] = Field(
+        description="Swell dissipation factor"
+    )
+
+    def cmd(self) -> str:
+        repr = f"{super().cmd()} ROGERS"
+        if self.cdsv is not None:
+            repr += f" cdsv={self.cdsv}"
+        if self.feswell is not None:
+            repr += f" feswell={self.feswell}"
+        return repr
+
+
+class ARDHUIN(SSWELL):
+    """Nonbreaking dissipation of Ardhuin et al. (2010).
+
+    References
+    ----------
+    Ardhuin, F., A. Rogers, A. Babanin, J. Filipot, J. Magne, A. Roland, and P. van
+    der Westhuysen, 2010: Semiempirical dissipation source functions for ocean waves.
+    Part I: Definition, calibration, and validation, J. Phys. Oceanogr., 40, 1917-1941.
+
+    """
+
+    model_type: Literal["ardhuin"] = Field(
+        default="ardhuin", description="Model type discriminator"
+    )
+    cdsv: Optional[float] = Field(
+        description="Coefficient related to laminar atmospheric boundary layer (SWAN default: 1.2)"
+    )
+
+    def cmd(self) -> str:
+        repr = f"{super().cmd()} ARDHUIN"
+        if self.cdsv is not None:
+            repr += f" cdsv={self.cdsv}"
+        return repr
+
+
+class ZIEGER(SSWELL):
+    """Nonbreaking dissipation of Young et al. (2013) updated by Zieger et al. (2015).
+
+    References
+    ----------
+    Zieger, S., A.V. Babanin, W. E. Rogers and I.R. Young, 2015: Observation-based
+    source terms in the third-generation wave model WAVEWATCH, Ocean Model., 96, 2-25.
+
+    Young, I.R., A.V.Babanin and S. Zieger, 2013: The Decay Rate of Ocean Swell
+    Observed by Altimeter, J. Phys. Oceanogr., 43, 3233-3233.
+
+    """
+
+    model_type: Literal["zieger"] = Field(
+        default="zieger", description="Model type discriminator"
+    )
+    b1: Optional[float] = Field(
+        description="Non-dimensional proportionality coefficient"
+    )
+
+    def cmd(self) -> str:
+        repr = f"{super().cmd()} ZIEGER"
+        if self.b1 is not None:
+            repr += f" b1={self.b1}"
+        return repr
+
+
+#======================================================================================
+# Physics group component
+#======================================================================================
 class PHYSICS(BaseComponent):
     """Physics component."""
 
@@ -190,6 +333,10 @@ class PHYSICS(BaseComponent):
         default=None,
         description="Swell dissipation specification",
         discriminator="model_type",
+    )
+    negatinp: Optional[NEGATINP] = Field(
+        default=None,
+        description="Negative wind input specification",
     )
     wcapping: WCAPKOMEN | WCAPAB | None = Field(
         default=None,
@@ -212,12 +359,28 @@ class PHYSICS(BaseComponent):
     def deactivate_physics(cls, values):
         return values
 
+    @validator("negatinp", pre=False)
+    def negatinp_only_with_zieger(cls, value, values):
+        """Log a warning if NEGATINP is used with a non-ZIEGER SSWELL."""
+        if values["sswell"] is None:
+            logger.warning(
+                "The negative wind input NEGATINP is only intended to use with the "
+                "swell dissipation SSWELL ZIEGER but no SSWELL has been specified."
+            )
+        elif values["sswell"].model_type != "zieger":
+            logger.warning(
+                "The negative wind input NEGATINP is only intended to use with the "
+                "swell dissipation SSWELL ZIEGER but "
+                f"SSWELL {values['sswell'].model_type.upper()} has been specified."
+            )
+        return value
+
     def cmd(self):
         repr = [self.gen.render()]
         if self.sswell is not None:
             repr += [f"{self.sswell.render()}"]
-        if self.sswell is not None and self.sswell.model_type == "zieger":
-            repr += [self.sswell.negatinp.render()]
+        if self.negatinp is not None:
+            repr += [self.negatinp.render()]
         if self.wcapping is not None:
             repr += [self.wcapping.render()]
         if self.quadrupl is not None:
