+++ rompy/swan/subcomponents/physics.py
@@ -280,150 +280,6 @@ class ST6C5(ST6C1):
     windscaling: Literal[35.0] = Field(default=35.0)
 
 
-class NEGATINP(BaseSubComponent):
-    """Nonbreaking dissipation of Young et al. (2013) updated by Zieger et al. (2015).
-
-    With this optional command the user activates negative wind input. This is intended
-    only for use with non-breaking swell dissipation SSWELL ZIEGER. Parameter `rdcoef`
-    is a fraction between 0 and 1, representing the strength of negative wind input. As
-    an example, with [rdcoef]=0.04, for a spectral bin that is opposed to the wind
-    direction, the wind input factor W(k, θ) is negative, and its magnitude is 4% of
-    the corresponding value of the spectral bin that is in the opposite direction
-    (i.e. in the wind direction). See Zieger et al. (2015) eq. 11, where a0 is their
-    notation for [rdcoef]. Default [rdcoef]=0.0 and `rdcoef=0.04` is recommended,
-    though as implied by Zieger et al. (2015), this value is not well-established, so
-    the user is encouraged to experiment with other values.
-
-    References
-    ----------
-    Zieger, S., A.V. Babanin, W. E. Rogers and I.R. Young, 2015: Observation-based
-    source terms in the third-generation wave model WAVEWATCH, Ocean Model., 96, 2-25.
-
-    """
-
-    model_type: Literal["negatinp"] = Field(
-        default="negatinp", description="Model type discriminator"
-    )
-    rdcoef: confloat(ge=0.0, le=1.0) = Field(
-        default=0.0,
-        description="Coefficient representing the strength of negative wind input",
-    )
-
-    def cmd(self) -> str:
-        return f"NEGATINP rdcoef={self.rdcoef}"
-
-
-#======================================================================================
-# SSWELL
-#======================================================================================
-class SSWELL(BaseSubComponent):
-    """Swell dissipation sub-component.
-
-    `SSWELL`
-
-    With this command the user can influence the type of swell dissipation included in
-    the computations. The Zieger option is intended for use with negative wind input
-    via the NEGATINP command. Zieger non-breaking dissipation follows the method used
-    in WAVEWATCH III version 4 and does not include the steepness-dependent swell
-    coefficient introduced in WAVEWATCH III version 5. As noted already, if
-    `GEN3 ST6...` command is used, this SSWELL command should be provided.
-
-    """
-
-    model_type: Literal["sswell"] = Field(
-        default="sswell", description="Model type discriminator"
-    )
-
-    def cmd(self) -> str:
-        return "SSWELL"
-
-
-class ROGERS(SSWELL):
-    """Nonbreaking dissipation of Rogers et al. (2012).
-
-    References
-    ----------
-    Rogers, W.E., A.V. Babanin and D.W. Wang, 2012: Observation-Consistent Input and
-    Whitecapping Dissipation in a Model for Wind-Generated Surface Waves: Description
-    and Simple Calculations, J. Atmos. Oceanic Technol., 29:9, 1329-1346.
-
-    """
-
-    model_type: Literal["rogers"] = Field(
-        default="rogers", description="Model type discriminator"
-    )
-    cdsv: Optional[float] = Field(
-        description="Coefficient related to laminar atmospheric boundary layer (SWAN default: 1.2)"
-    )
-    feswell: Optional[float] = Field(
-        description="Swell dissipation factor"
-    )
-
-    def cmd(self) -> str:
-        repr = f"{super().cmd()} ROGERS"
-        if self.cdsv is not None:
-            repr += f" cdsv={self.cdsv}"
-        if self.feswell is not None:
-            repr += f" feswell={self.feswell}"
-        return repr
-
-
-class ARDHUIN(SSWELL):
-    """Nonbreaking dissipation of Ardhuin et al. (2010).
-
-    References
-    ----------
-    Ardhuin, F., A. Rogers, A. Babanin, J. Filipot, J. Magne, A. Roland, and P. van
-    der Westhuysen, 2010: Semiempirical dissipation source functions for ocean waves.
-    Part I: Definition, calibration, and validation, J. Phys. Oceanogr., 40, 1917-1941.
-
-    """
-
-    model_type: Literal["ardhuin"] = Field(
-        default="ardhuin", description="Model type discriminator"
-    )
-    cdsv: Optional[float] = Field(
-        description="Coefficient related to laminar atmospheric boundary layer (SWAN default: 1.2)"
-    )
-
-    def cmd(self) -> str:
-        repr = f"{super().cmd()} ARDHUIN"
-        if self.cdsv is not None:
-            repr += f" cdsv={self.cdsv}"
-        return repr
-
-
-class ZIEGER(SSWELL):
-    """Nonbreaking dissipation of Young et al. (2013) updated by Zieger et al. (2015).
-
-    References
-    ----------
-    Zieger, S., A.V. Babanin, W. E. Rogers and I.R. Young, 2015: Observation-based
-    source terms in the third-generation wave model WAVEWATCH, Ocean Model., 96, 2-25.
-
-    Young, I.R., A.V.Babanin and S. Zieger, 2013: The Decay Rate of Ocean Swell
-    Observed by Altimeter, J. Phys. Oceanogr., 43, 3233-3233.
-
-    """
-
-    model_type: Literal["zieger"] = Field(
-        default="zieger", description="Model type discriminator"
-    )
-    b1: Optional[float] = Field(
-        description="Non-dimensional proportionality coefficient"
-    )
-    negatinp: NEGATINP = Field(
-        default=NEGATINP(),
-        description="Negative wind input specification",
-    )
-
-    def cmd(self) -> str:
-        repr = f"{super().cmd()} ZIEGER"
-        if self.b1 is not None:
-            repr += f" b1={self.b1}"
-        return repr
-
-
 #======================================================================================
 # WCAPPING
 #======================================================================================
